+++ NEWS.md
@@ -1,5 +1,8 @@
 # qualtRics (development version)
 
+- Add `fetch_description()` to download complete survey description metadata from v3 API endpoint (more up-to-date than older `metadata()`) thanks to @jmobrien (#207)
+- Warn users about possible incorrect results from API when `breakout_sets` and `label` are both FALSE
+
 # qualtRics 3.1.4
 
 - Add `fetch_distributions()` to access distribution data for a specific survey thanks to @dsen6644 (#169)